# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: "Описание спецификации всех API методов, используемых в online diffusion\
    \ сервисе"
  title: Online diffusion API
  version: 0.0.1
  x-logo:
    url: https://i.imgur.com/1kUh580.jpg
...
servers:
- description: "Локальный сервер, актуально для запуска на своем компьютере"
  url: http://localhost:8080
  x-last-modified: 1698183338768
security:
- JWTAuth: []
...
paths:
  /api/v1/auth/login:
    get:
      description: ""
      operationId: login
      parameters:
      - description: Почта пользователя
        explode: true
        in: query
        name: email
        required: true
        schema:
          example: foo@gmail.com
          format: email
          type: string
        style: form
      - description: Пароль пользователя
        explode: true
        in: query
        name: password
        required: true
        schema:
          example: querty123
          maxLength: 32
          minLength: 8
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthInfo'
          description: "все хорошо, авторизация прошла успешно"
        "400":
          description: логин или пароль неверные
      summary: Авторизация
      tags:
      - Auth
      x-accepts: application/json
      x-tags:
      - tag: Auth
  /api/v1/auth/refresh-token:
    put:
      description: ""
      operationId: refresh-token
      parameters:
      - description: "Токен, используемый для обновления авторизации, когда accessToken\
          \ кончился"
        explode: true
        in: query
        name: refreshToken
        required: true
        schema:
          example: "1234"
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthInfo'
          description: "все хорошо, регистрация прошла успешно, юзер должен подтвердить почту"
        "404":
          description: рефреш токен истек или невалиден
      summary: Обновление токена авторизации
      tags:
      - Auth
      x-accepts: application/json
      x-tags:
      - tag: Auth
  /api/v1/auth/register:
    post:
      description: ""
      operationId: register
      parameters:
      - description: Почта пользователя
        explode: true
        in: query
        name: email
        required: true
        schema:
          example: foo@gmail.com
          format: email
          type: string
        style: form
      - description: Имя пользователя
        explode: true
        in: query
        name: username
        required: true
        schema:
          example: user
          maxLength: 32
          minLength: 4
          type: string
        style: form
      - description: Пароль пользователя
        explode: true
        in: query
        name: password
        required: true
        schema:
          example: querty123
          maxLength: 32
          minLength: 8
          type: string
        style: form
      responses:
        "200":
          description: "все хорошо, регистрация прошла успешно, юзер должен подтве\
            рдить почту"
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/invalidData'
          description: "регистрация невозможна, так как какие-то из полей были нек\
            оректными"
        "409":
          description: "регистрация невозможна, так как некоторые поля удовлетворя\
            ют требованиям уникальности"
      summary: Регистрация на портале
      tags:
      - Auth
      x-accepts: application/json
      x-tags:
      - tag: Auth
  /api/v1/auth/recovery:
    post:
      description: ""
      operationId: recovery
      parameters:
      - description: Почта пользователя
        explode: true
        in: query
        name: email
        required: true
        schema:
          example: foo@gmail.com
          format: email
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/recoveryToken'
          description: "почта принята, код отправлен, если почта есть в базе"
        "400":
          description: "неккоректные данные, которые точно не являются почтовым ад\
            ресом"
      summary: Получение кода для восстановления на почту
      tags:
      - Auth
      x-accepts: application/json
      x-tags:
      - tag: Auth
  /api/v1/auth/confirmation:
    put:
      description: ""
      operationId: confirmation
      parameters:
      - description: Код для сброса пароля с электронной почты
        explode: true
        in: query
        name: code
        required: true
        schema:
          example: 123456
          format: int32
          type: integer
        style: form
      - description: "Токен, назначаемый сессии сброса пароля"
        explode: true
        in: query
        name: recoveryToken
        required: true
        schema:
          example: querty123
          type: string
        style: form
      responses:
        "200":
          content: {}
          description: "код правильный и соответсвует введеному recovery_token, мо\
            жно присылать новый пароль через метод Confirm_new_Password"
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/numOfTriesLeft'
          description: "присылается, когда такой recovery_token есть и он валидный\
            , однако код был введен неверно"
        "404":
          description: такого recovery_token нет на сервере в принципе
      summary: Подтверждение правильности кода
      tags:
      - Auth
      x-accepts: application/json
      x-tags:
      - tag: Auth
  /api/v1/auth/new-password:
    put:
      description: ""
      operationId: new-password
      parameters:
      - description: Код для сброса пароля с электронной почты
        explode: true
        in: query
        name: code
        required: true
        schema:
          example: 123456
          format: int32
          type: integer
        style: form
      - description: "Токен, назначаемый сессии сброса пароля"
        explode: true
        in: query
        name: recoveryToken
        required: true
        schema:
          example: querty123
          type: string
        style: form
      responses:
        "200":
          content: {}
          description: токен активирован и новый пароль успешно принят
        "400":
          description: "такой токен есть, однако он еще не был активирован присыла\
            нием правильного кода в методе Confirm_Password_Code"
        "404":
          description: такого recovery_token нет на сервере в принципе
      summary: Ввод нового пароля
      tags:
      - Auth
      x-accepts: application/json
      x-tags:
      - tag: Auth
...
components:
...
  schemas:
...
    invalidData:
      description: показывает какие именно поля были некорректными
      enum:
      - email
      - username
      - password
      type: string
    recoveryToken:
      description: "Токен, назначаемый сессии сброса пароля"
      type: string
    numOfTriesLeft:
      description: приходит в случае ввода неправильного кода для токена. Указывает
        сколько попыток еще осталось
      format: int32
      type: integer
...
    AuthInfo:
      description: "Ответная информация, содержащая JWT токены пользователя"
      example:
        accessToken: accessToken
        refreshToken: refreshToken
      properties:
        refreshToken:
          description: "Токен, используемый для обновления авторизации, когда accessToken\
            \ кончился"
          type: string
        accessToken:
          description: JWT токен авторизированного пользователя
          type: string
      required:
      - accessToken
      - refreshToken
      type: object
  securitySchemes:
    JWTAuth:
      bearerFormat: JWT
      description: "Example: \n> Authorization: Bearer <token>"
      scheme: bearer
      type: http
      x-last-modified: 1698183246392